requirements.txt
flask
psycopg2-binary
python-dotenv
flask_sqlalchemy
gunicorn==21.2.0
requests
openai
openpyxl
pandas
sqlalchemy
flask_cors